refactor minter_test.go blockchain
