+++ core/minter/minter_test.go
@@ -32,7 +32,7 @@ import (
 	"time"
 )
 
-func initTestNode(t *testing.T) (*Blockchain, *rpc.Local, *privval.FilePV, func()) {
+func initTestNode(t *testing.T, initialHeight int64) (*Blockchain, *rpc.Local, *privval.FilePV, func()) {
 	storage := utils.NewStorage(t.TempDir(), "")
 	minterCfg := config.GetConfig(storage.GetMinterHome())
 	logger := log.NewLogger(minterCfg)
@@ -67,7 +67,7 @@ func initTestNode(t *testing.T) (*Blockchain, *rpc.Local, *privval.FilePV, func(
 		pv,
 		nodeKey,
 		proxy.NewLocalClientCreator(app),
-		getTestGenesis(pv, storage.GetMinterHome()),
+		getTestGenesis(pv, storage.GetMinterHome(), initialHeight),
 		tmNode.DefaultDBProvider,
 		tmNode.DefaultMetricsProvider(cfg.Instrumentation),
 		logger,
@@ -113,12 +113,69 @@ func initTestNode(t *testing.T) (*Blockchain, *rpc.Local, *privval.FilePV, func(
 }
 
 func TestBlockchain_Run(t *testing.T) {
-	_, _, _, cancel := initTestNode(t)
+	_, _, _, cancel := initTestNode(t, 0)
 	cancel()
 }
 
+func TestBlockchain_InitialBlockHeight(t *testing.T) {
+	blockchain, tmCli, _, cancel := initTestNode(t, 100)
+	defer cancel()
+
+	value := helpers.BipToPip(big.NewInt(10))
+	to := types.Address([20]byte{1})
+
+	data := transaction.SendData{
+		Coin:  types.GetBaseCoinID(),
+		To:    to,
+		Value: value,
+	}
+
+	encodedData, err := rlp.EncodeToBytes(data)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	nonce := uint64(1)
+	tx := transaction.Transaction{
+		Nonce:         nonce,
+		ChainID:       types.CurrentChainID,
+		GasPrice:      1,
+		GasCoin:       types.GetBaseCoinID(),
+		Type:          transaction.TypeSend,
+		Data:          encodedData,
+		SignatureType: transaction.SigTypeSingle,
+	}
+	nonce++
+
+	if err := tx.Sign(getPrivateKey()); err != nil {
+		t.Fatal(err)
+	}
+
+	txBytes, err := tx.Serialize()
+	if err != nil {
+		t.Fatalf("Failed: %s", err.Error())
+	}
+
+	res, err := tmCli.BroadcastTxCommit(context.Background(), txBytes)
+	if err != nil {
+		t.Fatalf("Failed: %s", err.Error())
+	}
+
+	time.Sleep(time.Second)
+
+	resultTx, err := tmCli.Tx(context.Background(), res.Hash.Bytes(), false)
+	if err != nil {
+		t.Fatalf("Failed: %s", err.Error())
+	}
+
+	_, err = blockchain.GetStateForHeight(uint64(resultTx.Height - 1))
+	if err != nil {
+		t.Fatalf("Failed: %s", err.Error())
+	}
+}
+
 func TestBlockchain_Height(t *testing.T) {
-	blockchain, tmCli, _, cancel := initTestNode(t)
+	blockchain, tmCli, _, cancel := initTestNode(t, 100)
 	defer cancel()
 
 	blocks, err := tmCli.Subscribe(context.Background(), "test-client", "tm.event = 'NewBlock'")
@@ -147,7 +204,7 @@ func TestBlockchain_Height(t *testing.T) {
 }
 
 func TestBlockchain_SetStatisticData(t *testing.T) {
-	blockchain, tmCli, _, cancel := initTestNode(t)
+	blockchain, tmCli, _, cancel := initTestNode(t, 0)
 	defer cancel()
 
 	ch := make(chan struct{})
@@ -190,7 +247,7 @@ func TestBlockchain_SetStatisticData(t *testing.T) {
 }
 
 func TestBlockchain_IsApplicationHalted(t *testing.T) {
-	blockchain, tmCli, pv, cancel := initTestNode(t)
+	blockchain, tmCli, pv, cancel := initTestNode(t, 0)
 	defer cancel()
 	data := transaction.SetHaltBlockData{
 		PubKey: types.BytesToPubkey(pv.Key.PubKey.Bytes()[:]),
@@ -260,7 +317,7 @@ func TestBlockchain_IsApplicationHalted(t *testing.T) {
 }
 
 func TestBlockchain_GetStateForHeightAndDeleteStateVersions(t *testing.T) {
-	blockchain, tmCli, _, cancel := initTestNode(t)
+	blockchain, tmCli, _, cancel := initTestNode(t, 100)
 	defer cancel()
 
 	symbol := types.StrToCoinSymbol("AAA123")
@@ -339,7 +396,7 @@ func TestBlockchain_GetStateForHeightAndDeleteStateVersions(t *testing.T) {
 }
 
 func TestBlockchain_SendTx(t *testing.T) {
-	blockchain, tmCli, _, cancel := initTestNode(t)
+	blockchain, tmCli, _, cancel := initTestNode(t, 0)
 	defer cancel()
 
 	value := helpers.BipToPip(big.NewInt(10))
@@ -411,7 +468,7 @@ func TestBlockchain_SendTx(t *testing.T) {
 }
 
 func TestBlockchain_FrozenFunds(t *testing.T) {
-	blockchain, tmCli, pv, cancel := initTestNode(t)
+	blockchain, tmCli, pv, cancel := initTestNode(t, 0)
 	defer cancel()
 
 	targetHeight := uint64(10)
@@ -459,7 +516,7 @@ func TestBlockchain_FrozenFunds(t *testing.T) {
 }
 
 func TestBlockchain_RecalculateStakes_andRemoveValidator(t *testing.T) {
-	blockchain, tmCli, _, cancel := initTestNode(t)
+	blockchain, tmCli, _, cancel := initTestNode(t, 0)
 	defer cancel()
 
 	txs, err := tmCli.Subscribe(context.Background(), "test-client", "tm.event = 'Tx'")
@@ -716,7 +773,7 @@ func TestBlockchain_RecalculateStakes_andRemoveValidator(t *testing.T) {
 }
 
 func TestStopNetworkByHaltBlocks(t *testing.T) {
-	blockchain, _, _, cancel := initTestNode(t)
+	blockchain, _, _, cancel := initTestNode(t, 0)
 	cancel()
 
 	haltHeight := uint64(50)
@@ -796,12 +853,11 @@ func makeTestValidatorsAndCandidates(pubkeys []string, stake *big.Int) ([]types.
 	return vals, cands
 }
 
-func getTestGenesis(pv *privval.FilePV, home string) func() (*types2.GenesisDoc, error) {
+func getTestGenesis(pv *privval.FilePV, home string, initialState int64) func() (*types2.GenesisDoc, error) {
 	return func() (*types2.GenesisDoc, error) {
 		validators, candidates := makeTestValidatorsAndCandidates([]string{string(pv.Key.PubKey.Bytes()[:])}, helpers.BipToPip(big.NewInt(12444011)))
 
 		appState := types.AppState{
-			// StartHeight: 100, // FIXME
 			TotalSlashed: "0",
 			Accounts: []types.Account{
 				{
@@ -824,10 +880,11 @@ func getTestGenesis(pv *privval.FilePV, home string) func() (*types2.GenesisDoc,
 		}
 
 		genesisDoc := types2.GenesisDoc{
-			ChainID:     "minter-test-network",
-			GenesisTime: time.Now(),
-			AppHash:     nil,
-			AppState:    json.RawMessage(appStateJSON),
+			ChainID:       "minter-test-network",
+			InitialHeight: initialState,
+			GenesisTime:   time.Now(),
+			AppHash:       nil,
+			AppState:      json.RawMessage(appStateJSON),
 		}
 
 		err = genesisDoc.ValidateAndComplete()
