+++ core/types/appstate.go
@@ -9,7 +9,6 @@ import (
 
 type AppState struct {
 	Note                string       `json:"note"`
-	StartHeight         uint64       `json:"start_height"`
 	Validators          []Validator  `json:"validators,omitempty"`
 	Candidates          []Candidate  `json:"candidates,omitempty"`
 	BlockListCandidates []Pubkey     `json:"block_list_candidates,omitempty"`
